sx1301: Refactor GNSS handling.
This aligns the SX1301 GNSS implementation with the SX1302.

+++ chirpstack-concentratord-sx1301/src/config/mod.rs
@@ -70,7 +70,7 @@ pub struct Beacon {
 pub struct Location {
     pub latitude: f64,
     pub longitude: f64,
-    pub altitude: i16,
+    pub altitude: f32,
 }
 
 #[derive(Serialize, Deserialize)]
